ISSUE #2754 Unit and processor tests added for favourites

+++ backend/src/v5/processors/teamspaces/projects/federations/federations.js
@@ -16,16 +16,29 @@
  */
 
  const { appendFavourites, deleteFavourites } = require('../../../../models/users');
+ const { getProjectById } = require('../../../../models/projects');
 
  const Federations = {};
  
  Federations.appendFavourites = async (username,teamspace, favouritesToAdd) => {
-     return await appendFavourites(username, teamspace, favouritesToAdd);
+    const {  models } = await getProjectById(teamspace, project, { permissions: 1, models: 1 });
+	
+	if(favouritesToAdd.every(i=>models.includes(i))) {
+		appendFavourites(username, teamspace, favouritesToAdd);
+	}
+
+	return;
  };
  
  Federations.deleteFavourites = async (username,teamspace,favouritesToRemove) => {	
-     return deleteFavourites(username, teamspace, favouritesToRemove);
- };
+    const {  models } = await getProjectById(teamspace, project, { permissions: 1, models: 1 });
+	
+	if(favouritesToRemove.every(i=>models.includes(i))) {
+		deleteFavourites(username, teamspace, favouritesToRemove);
+	}
+
+	return; 
+};
  
  module.exports = Federations;
